feat(scaners): nmap scanner

See merge request dev/main/extrnode/scanner-api-back-golang!3

+++ internal/api/api.go
@@ -80,7 +80,7 @@ func NewAPI(cfg config.Config) (*api, error) {
 	}
 
 	a := &api{
-		port:    uint64(cfg.API.Port),
+		port:    cfg.API.Port,
 		router:  echo.New(),
 		storage: s,
 		cache:   cache.New(cacheTTL, cacheTTL),
